+++ FinanceTrack.Finance/src/FinanceTrack.Finance.Core/Interfaces/ITelegramSender.cs
@@ -1,0 +1,12 @@
+namespace FinanceTrack.Finance.Core.Interfaces;
+
+public interface ITelegramSender
+{
+    /// <summary>
+    /// Возвращает true, если бот сконфигурирован (токен задан).
+    /// Используется для пропуска отправки без выброса исключения.
+    /// </summary>
+    bool IsConfigured { get; }
+
+    Task SendMessageAsync(long chatId, string text, CancellationToken cancel = default);
+}
